Start extract clpp.h ; new Makefile
