+++ Makefile
@@ -1,24 +1,42 @@
-HEADERS = Background.h Pm1Plan.h GmpUtil.h Args.h checkpoint.h clwrap.h common.h kernel.h state.h timeutil.h tinycl.h Worktodo.h Gpu.h Signal.h FFTConfig.h version.h version.inc
+SRCS = Pm1Plan.cpp GmpUtil.cpp Worktodo.cpp common.cpp main.cpp Gpu.cpp clwrap.cpp Task.cpp checkpoint.cpp timeutil.cpp Args.cpp state.cpp Signal.cpp FFTConfig.cpp clpp.cpp
 
-SRCS = Pm1Plan.cpp GmpUtil.cpp Worktodo.cpp common.cpp main.cpp Gpu.cpp clwrap.cpp Task.cpp checkpoint.cpp timeutil.cpp Args.cpp state.cpp Signal.cpp FFTConfig.cpp
+CXXFLAGS = -Wall -O2 -std=c++17
 
-# Edit the path in -L below if needed, to the folder containing OpenCL.dll on Windows or libOpenCL.so on UNIX.
-# The included lib paths are for ROCm, AMDGPU-pro/Linux or MSYS-2/Windows.
 LIBPATH = -L/opt/rocm/opencl/lib/x86_64 -L/opt/amdgpu-pro/lib/x86_64-linux-gnu -L/c/Windows/System32 -L.
 
-#-fsanitize=leak
+LDFLAGS = -lOpenCL -lgmp -lstdc++fs -pthread ${LIBPATH}
 
-BUILD = g++ -Wall -O2 -std=c++17 -Wall ${SRCS} -o $@ -lOpenCL -lgmp -lstdc++fs -pthread ${LIBPATH}
+LINK = $(CXX) -o $@ ${OBJS} ${LDFLAGS}
 
-gpuowl: ${HEADERS} ${SRCS}
-	${BUILD}
 
-gpuowl-win: ${HEADERS} ${SRCS}
-	${BUILD} -static
+OBJS = $(SRCS:%.cpp=%.o)
+DEPDIR := .d
+$(shell mkdir -p $(DEPDIR) >/dev/null)
+DEPFLAGS = -MT $@ -MMD -MP -MF $(DEPDIR)/$*.Td
+COMPILE.cc = $(CXX) $(DEPFLAGS) $(CXXFLAGS) $(CPPFLAGS) $(TARGET_ARCH) -c
+POSTCOMPILE = @mv -f $(DEPDIR)/$*.Td $(DEPDIR)/$*.d && touch $@
+
+
+gpuowl: ${OBJS}
+	${LINK}
+
+gpuowl-win: ${OBJS}
+	{LINK} -static
 	strip $@
 
+%.o : %.cpp
+%.o : %.cpp $(DEPDIR)/%.d
+	$(COMPILE.cc) $(OUTPUT_OPTION) $<
+	$(POSTCOMPILE)
+
+$(DEPDIR)/%.d: ;
+.PRECIOUS: $(DEPDIR)/%.d
+
 version.inc: FORCE
-	echo \"`git describe --long --dirty --always`\" > version.inc
+	echo \"`git describe --long --dirty --always`\" > version.new
+	diff -q -N version.new version.inc >/dev/null || mv version.new version.inc
 	echo Version: `cat version.inc`
 
 FORCE:
+
+include $(wildcard $(patsubst %,$(DEPDIR)/%.d,$(basename $(SRCS))))
