+++ clwrap.h
@@ -3,14 +3,15 @@
 #pragma once
 
 #include "tinycl.h"
-#include "common.h"
+// #include "common.h"
 
 #include <string>
 #include <vector>
-#include <memory>
 #include <cassert>
+#include <memory>
 
 using cl_queue = cl_command_queue;
+
 
 void release(cl_context context);
 void release(cl_kernel k);
@@ -37,28 +38,6 @@ template<typename T> using Holder = std::unique_ptr<T, Deleter<T> >;
 using Context = std::unique_ptr<cl_context>;
 using QueueHolder = std::unique_ptr<cl_queue>;
 
-template<typename T>
-class Buffer : public std::unique_ptr<cl_mem> {
-  size_t _size{};
-  
-public:
-  Buffer() = default;
-  
-  Buffer(cl_context context, unsigned kind, size_t size, const T* ptr = nullptr)
-    : std::unique_ptr<cl_mem>{_makeBuf(context, kind, size * sizeof(T), ptr)}
-    , _size(size)
-  {}
-
-  Buffer(const Context& context, unsigned kind, size_t size, const T* ptr = nullptr)
-    : Buffer(context.get(), kind, size, ptr)
-  {}
-  
-  Buffer(cl_context context, unsigned kind, const std::vector<T>& vect)
-    : Buffer(context, kind, vect.size(), vect.data())
-  {}
-
-  size_t size() const { return _size; }
-};
 
 const unsigned BUF_CONST = CL_MEM_READ_ONLY | CL_MEM_COPY_HOST_PTR | CL_MEM_HOST_NO_ACCESS;
 const unsigned BUF_RW    = CL_MEM_READ_WRITE | CL_MEM_HOST_NO_ACCESS;
@@ -88,20 +67,10 @@ cl_kernel makeKernel(cl_program program, const char *name);
 template<typename T>
 void setArg(cl_kernel k, int pos, const T &value) { CHECK1(clSetKernelArg(k, pos, sizeof(value), &value)); }
 
-template<typename T>
-void setArg(cl_kernel k, int pos, const Buffer<T>& buf) { setArg(k, pos, buf.get()); }
-
-// template<> void setArg<void*>(cl_kernel k, int pos, void* const &value) {CHECK(clSetKernelArgSVMPointer(k, pos, value));}
-
 template<>
 void setArg<int>(cl_kernel k, int pos, const int &value);
 
-// Special-case Buffer argument: pass the wrapped cl_mem.
-template<typename T>
-void setArg(cl_kernel k, int pos, const Buffer<T>& buf);
-
 cl_mem _makeBuf(cl_context context, unsigned kind, size_t size, const void *ptr = 0);
-// cl_mem makeBuf(Context &context, unsigned kind, size_t size, const void *ptr = 0);
 cl_queue makeQueue(cl_device_id d, cl_context c);
 
 void flush( cl_queue q);
@@ -124,53 +93,6 @@ int getKernelNumArgs(cl_kernel k);
 int getWorkGroupSize(cl_kernel k, cl_device_id device, const char *name);
 std::string getKernelArgName(cl_kernel k, int pos);
 
-class Queue : public QueueHolder {
-public:
-  explicit Queue(cl_queue queue) : QueueHolder(queue) {}
-
-  template<typename T> vector<T> read(const Buffer<T>& buf, size_t sizeOrFull = 0) {
-    auto size = sizeOrFull ? sizeOrFull : buf.size();
-    assert(size <= buf.size());
-    vector<T> ret(size);
-    ::read(get(), true, buf.get(), size * sizeof(T), ret.data());
-    return ret;
-  }
-
-  template<typename T> void readAsync(const Buffer<T>& buf, vector<T>& out, size_t sizeOrFull = 0) {
-    auto size = sizeOrFull ? sizeOrFull : buf.size();
-    assert(size <= buf.size());
-    out.resize(size);
-    ::read(get(), false, buf.get(), size * sizeof(T), out.data());
-  }
-    
-  template<typename T> void write(Buffer<T>& buf, const vector<T> &vect) {
-    assert(vect.size() <= buf.size());
-    ::write(get(), true, buf.get(), vect.size() * sizeof(T), vect.data());
-  }
-
-  template<typename T> void writeAsync(Buffer<T>& buf, const vector<T> &vect) {
-    assert(vect.size() <= buf.size());
-    ::write(get(), false, buf.get(), vect.size() * sizeof(T), vect.data());
-  }
-
-  template<typename T> void copyFromTo(const Buffer<T>& src, Buffer<T>& dst) {
-    assert(src.size() <= dst.size());
-    copyBuf(get(), src.get(), dst.get(), src.size() * sizeof(T));
-  }
-  
-  void run(cl_kernel kernel, size_t groupSize, size_t workSize, const string &name) {
-    ::run(get(), kernel, groupSize, workSize, name);
-  }
-
-  void finish() { ::finish(get()); }
-
-  template<typename T> void zero(Buffer<T>& buf, size_t sizeOrFull = 0) {
-    auto size = sizeOrFull ? sizeOrFull : buf.size();
-    assert(size <= buf.size());
-    T zero = 0;
-    fillBuf(get(), buf.get(), &zero, sizeof(T), size * sizeof(T));
-  }
-};
 
 cl_device_id getDevice(int argsDevId);
