+++ kernel.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "clwrap.h"
+#include "clpp.h"
 #include "timeutil.h"
 #include "common.h"
